updated to only prod run

+++ src/cronjobs/FinancialAnalytics.cron.ts
@@ -277,6 +277,10 @@ async function sendDiscordNotice() {
 }
 
 export async function doFinancialAnalytics() {
+    if (Config.ENV.NODE_ENV !== 'production') {
+        LogManager.debug('CronJobs: doFinancialAnalytics() will only execute in production.');
+        return;
+    }
     LogManager.info('Starting Financial Analytics');
     const deleteOldScans = await BotDB.delete(faScans).where(
         lt(faScans.createdAt, new Date(Date.now() - 7 * 24 * 60 * 60 * 1000)),
